assignBlocks dist cutoff
